move home dashboard tests into HomeEnrollmentsDashboard.test.tsx
The 16 home tests now live alongside the component they cover and
render <HomeEnrollmentsDashboard /> directly instead of going through
the EnrollmentDisplay dispatcher. EnrollmentDisplay.test.tsx keeps
only the program tests (will move to ProgramEnrollmentDashboard.test.tsx
in Phase 4).

Trims dead imports from EnrollmentDisplay.test.tsx (act, setupEnrollments)
and from the new file (waitFor, mockAxiosInstance, invariant).
